+++ base/SlashCommand.js
@@ -1,0 +1,15 @@
+class SlashCommand {
+  constructor (client, {
+    name = null,
+    description = "No description provided",
+    options = [],
+    defaultPermissions = true,
+    guildOnly = false
+  }) {
+    this.client = client;
+    this.commandData = {name, description, options, defaultPermissions};
+    this.guildOnly = guildOnly;
+  }
+}
+
+module.exports = SlashCommand;