+++ commands/check.js
@@ -3,10 +3,11 @@ const { DateTime } = require("luxon");
 
 exports.run = async (client, message, args, level) => {
   if (!args[0]) return message.channel.send(`Improper usage. Use \`${client.getSettings(message.guild.id).prefix}help check\` for assistance.`);
+  var embed;
   switch (args[0].toLowerCase()) {
     case "horny":
       if (!message.mentions.users.first()) var users = [message.author];
-      else var users = message.mentions.users.array();
+      else var users = message.mentions.users.values();
       var fields = [];
       var i = 0;
 
@@ -32,14 +33,15 @@ exports.run = async (client, message, args, level) => {
         }
         i++;
       }
-      message.channel.send({"embed": {
+      embed = {
         "fields": fields, 
         "color": client.colorInt("#ff0000")
-      }});
+      };
+      message.channel.send({"embeds": [embed]});
       break;
     case "abuse":
       if (!message.mentions.users.first()) var users = [message.guild.owner];
-      else var users = message.mentions.users.array();
+      else var users = message.mentions.users.values();
       var fields = [];
       var i = 0;
 
@@ -66,14 +68,15 @@ exports.run = async (client, message, args, level) => {
         }
         i++;
       }
-      message.channel.send({"embed": {
+      embed = {
         "fields": fields, 
         "color": client.colorInt("#ff0000")
-      }});
+      };
+      message.channel.send({"embeds": [embed]});
       break;
     case "gratitude":
       if (!message.mentions.users.first()) var users = [message.guild.owner];
-      else var users = message.mentions.users.array();
+      else var users = message.mentions.users.values();
       var fields = [];
       var i = 0;
 
@@ -100,10 +103,12 @@ exports.run = async (client, message, args, level) => {
         }
         i++;
       }
-      message.channel.send({"embed": {
+
+      embed = {
         "fields": fields, 
         "color": client.colorInt("#00ff00")
-      }});
+      };
+      message.channel.send({"embeds": [embed]});
       break;
     default:
       return message.channel.send(`Improper usage. Use \`${client.getSettings(message.guild.id).prefix}help check\` for assistance.`);