+++ commands/gratitude.js
@@ -10,10 +10,11 @@ exports.run = async (client, message, args, level) => {
   const profile = await Special.findById(owner.id) || {"gratitude": 0};
   const res = await Special.findByIdAndUpdate(owner.id, {"gratitude": profile.gratitude ? profile.gratitude + 1 : 1}, {upsert: true, new: true});
 
-  message.channel.send({"embed": {
+  const embed = {
     "description": `Thank you <@${owner.id}> for everything you do! <@${owner.id}> has been thanked ${res.gratitude} time${res.gratitude > 1 ? "s" : "" }`,
     "color": client.colorInt("#00ff00")
-  }});
+  };
+  message.channel.send({"embeds": [embed]});
 };
 
 exports.conf = {