+++ commands/list.js
@@ -30,11 +30,13 @@ exports.run = async (client, message, args, level) => {
         fieldarray[i] = {"name": "AniList", "value": `[${result.anilist}](https://anilist.co/user/${result.kitsu})`};
         i++;
       }
-      message.channel.send({"embed": {
+
+      const embed = {
         "title": `${usermention.username}'s Anime Lists`,
         "thumbnail": { "url": usermention.avatarURL() || usermention.defaultAvatarURL() },
         "fields": fieldarray
-      }});
+      };
+      message.channel.send({"embeds": [embed]});
     });
   } else {
     // Add mode - Add a list to the user's profile