+++ commands/source.js
@@ -14,7 +14,8 @@ exports.run = async (client, message, args, level) => {
   }
 
   var results = await saucenao(img);
-  message.channel.send({"embed": {
+
+  const embed = {
     "title": results[0].raw.data.title || `Image from ${results[0].site}`,
     "url": results[0].url,
     "color": client.colorInt("#1d1d1d"),
@@ -23,7 +24,8 @@ exports.run = async (client, message, args, level) => {
       { "name": "Similarity", "value": `${results[0].similarity}` },
       { "name": results[0].raw.data.anidb_aid ? "Anime" : "Artist", "value": `${results[0].raw.data.anidb_aid ? results[0].raw.data.source : results[0].raw.data.creator || `${results[0].raw.data.member_name} (${results[0].raw.data.member_id})`}` }
     ]
-  }});
+  };
+  message.channel.send({"embeds": [embed]});
   client.logger.log(`Result from ${results[0].site} found for ${img}`);
 };