+++ commands/stats.js
@@ -1,9 +1,10 @@
 const { version } = require("discord.js");
-const moment = require("moment");
-require("moment-duration-format");
+//const moment = require("moment");
+//require("moment-duration-format");
 
 exports.run = (client, message, args, level) => {
-  const duration = moment.duration(client.uptime).format(" D [days], H [hrs], m [mins], s [secs]");
+  //const duration = moment.duration(client.uptime).format(" D [days], H [hrs], m [mins], s [secs]");
+  const duration = "N/A";
   message.channel.send(`= STATISTICS =
 • Mem Usage  :: ${(process.memoryUsage().heapUsed / 1024 / 1024).toFixed(2)} MB
 • Uptime     :: ${duration}