+++ commands/vnsearch.js
@@ -1,6 +1,6 @@
 // Search for a visual novel on VNDB
 const VNDB = require("vndb-api");
-const moment = require("moment");
+const { DateTime } = require("luxon");
 
 exports.run = async (client, message, args, level) => {
   if (!args[0]) return message.channel.send("Please specify a visual novel title.");
@@ -35,15 +35,17 @@ exports.run = async (client, message, args, level) => {
     }
     fieldarry[i] = {
       "name": vnresult.title,
-      "value": `Release Date: ${moment(vnresult.released).format("MMM D[,] YYYY")}\nLanguages: ${langs.join(" ")}\nPlatforms: ${plats.join(", ")}\n[VNDB.org](https://vndb.org/v${vnresult.id})`
+      "value": `Release Date: ${DateTime.fromISO(vnresult.released).toFormat("MMM d',' yyyy")}\nLanguages: ${langs.join(" ")}\nPlatforms: ${plats.join(", ")}\n[VNDB.org](https://vndb.org/v${vnresult.id})`
     };
   }
-  message.channel.send({"embed": {
+
+  const embed = {
     "title": "Search Results",
     "description": "\u200b",
     "color": client.colorInt("#071c30"),
     "fields": fieldarry
-  }});
+  };
+  message.channel.send({"embeds": [embed]});
   client.logger.log(`Result found for search term "${vnname}"`);
 };