+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 6.0 - v13 Update (2021-08-09)
+- Upgrades for `discord.js` v13
+- Implementing new Client builder
+- Node/NPM updates
+- Package updates
+
 ## 5.3 - I'm Too Dependent (2021-07-12)
 - Replacing `enmap` dependency with `mongoose`
 - Replacing `moment` dependency with `luxon`
-- Package Updates
+- Package updates
 
 ## 5.2 - The Quadfecta (2020-04-28)
 - Added custom module for WLN Updates API requests